Added subclasses to level template (very basic) - vchadha
